crud admin - kelola user

+++ application/views/admin/kelola_user.php
@@ -57,7 +57,7 @@ $this->load->view('inc/sidebar');
                     <td>$row[name]</td>
                     <td><center>
                       <a class='btn btn-fill btn-warning btn-xs' data-toggle='tooltip' data-placement='bottom' title='Ubah' href='".base_url()."kelolauser/ubah/$row[id_du]'><span class='glyphicon glyphicon-edit'></span></a>
-                      <a class='btn btn-fill btn-danger btn-xs' data-toggle='tooltip' data-placement='bottom' title='Hapus' href='".base_url()."kelolauser/hapus/$row[id_du]'><span class='fa fa-trash-o'></span></a>
+                      <a class='btn btn-fill btn-danger btn-xs' data-toggle='tooltip' data-placement='bottom' title='Hapus' href='".base_url()."kelolauser/hapus/$row[id_du]' onclick=\"return confirm('Apa anda yakin untuk hapus Data ini?')\"><span class='fa fa-trash-o'></span></a>
                     </tr>";
                 $no++;
               }
